Add r8 impl (#6)
* Add audio and keyboard

* Fix timers

* Fix typo

* Update GH Action config

+++ r8lib/src/operations/op00ee.rs
@@ -21,7 +21,7 @@ impl Operation for Op00ee {
     fn exec(&self, machine: &mut Machine) -> OperationResult {
         debug!("op_00ee");
 
-        machine.pc = machine.stack[machine.sp as usize] as usize;
+        machine.pc = machine.stack[machine.sp] as usize;
         machine.sp -= 1;
 
         OperationResult::Next
